% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{sc_case}
\alias{sc_case}
\title{Fetch Supreme Court Docket Details from the Oyez API}
\usage{
sc_case(term, docket_number)
}
\arguments{
\item{term}{Integer. The court term year (e.g., 2022).}

\item{docket_number}{Character. The case docket number (e.g., "20-1199").}
}
\value{
A `tibble::tibble` containing case metadata
}
\description{
Fetch Supreme Court Docket Details from the Oyez API
}
\examples{
sffa <- sc_case(2022, "20-1199")
}
