% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{sc_cases}
\alias{sc_cases}
\title{Fetch Supreme Court Cases from Oyez API}
\usage{
sc_cases(term = NULL, per_page = 0)
}
\arguments{
\item{term}{Integer or character. Supreme court term year (e.g., 2022, "2022)}

\item{per_page}{Integer. Number of results per page. Set to 0 to fetch all cases in term. Default is 0}
}
\value{
A tidy tibble of Supreme Court cases.
}
\description{
Fetch Supreme Court Cases from Oyez API
}
\examples{
sc_cases(term = 2022)
}
